fix(app): split create helpers to satisfy lint max-lines
